% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_product_sensitivity}
\alias{calculate_product_sensitivity}
\title{functions to make ultrametric trees
callable loci correction
Function to calculate the product of (1 - sensitivity) for each sample in a node
Split the 'samples' string by '|'
Retrieve sensitivities for these samples
Calculate the product of (1 - sensitivity)}
\usage{
calculate_product_sensitivity(samples, df_sensitivity)
}
\description{
functions to make ultrametric trees
callable loci correction
Function to calculate the product of (1 - sensitivity) for each sample in a node
Split the 'samples' string by '|'
Retrieve sensitivities for these samples
Calculate the product of (1 - sensitivity)
}
