% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correct_branches}
\alias{correct_branches}
\title{function to take a tree and correct the branch lengths based on the product sensitivity
requires samples (samples column from output of cellphywrapper object: tree@data$samples .
requires a CALLABLE df (containing columns SAMPLES and CALLABLE (in nr of bases)
Function to calculate the product of (1 - sensitivity) for each sample in a node
get sensitivity df by taking fraction of total possible callable loci
calculate the sensitivity per node/branch using the calculate_product_sensitivity function
correct the branch lengths
store also in the phylo object, required for downstream analyses}
\usage{
correct_branches(tree, callable_df)
}
\description{
function to take a tree and correct the branch lengths based on the product sensitivity
requires samples (samples column from output of cellphywrapper object: tree@data$samples .
requires a CALLABLE df (containing columns SAMPLES and CALLABLE (in nr of bases)
Function to calculate the product of (1 - sensitivity) for each sample in a node
get sensitivity df by taking fraction of total possible callable loci
calculate the sensitivity per node/branch using the calculate_product_sensitivity function
correct the branch lengths
store also in the phylo object, required for downstream analyses
}
