% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_wholelife_tree}
\alias{create_wholelife_tree}
\title{full wrapper for ultrametric tree building from 'correct_branches' onwards.
uses output of correct_branches}
\usage{
create_wholelife_tree(corr_tree, age_HSC_at_FU, mutload_birth = 79)
}
\description{
full wrapper for ultrametric tree building from 'correct_branches' onwards.
uses output of correct_branches
}
