% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_interval_branch}
\alias{get_interval_branch}
\title{retrieving the distances from root to branch end
gives output for upper, lower, and estimated nr of mutations at birth
input: output of correct_branches
plotting interval of branch timing (based on different possible mutation loads at birth)
get trees with upper and lower boundary of HSC load at birth
get node
get values}
\usage{
get_interval_branch(
  CLcorrTree,
  branch_of_interest,
  age_HSC_at_FU,
  ave_mutload_birth = 79,
  lower_mutload_birth = 50,
  higher_mutload_birth = 108.9
)
}
\description{
retrieving the distances from root to branch end
gives output for upper, lower, and estimated nr of mutations at birth
input: output of correct_branches
plotting interval of branch timing (based on different possible mutation loads at birth)
get trees with upper and lower boundary of HSC load at birth
get node
get values
}
