% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{length.normalise}
\alias{length.normalise}
\title{Call as recursive function}
\usage{
\method{length}{normalise}(orig.tree, new.tree, curr.node, remaining.stick)
}
\description{
Call as recursive function
}
