% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.ultrametric.tree}
\alias{make.ultrametric.tree}
\title{ultrametric making ##' FROM FABRE ET AL 2022 NATURE https://doi.org/10.1038/s41586-022-04785-z}
\usage{
make.ultrametric.tree(tree)
}
\description{
ultrametric making ##' FROM FABRE ET AL 2022 NATURE https://doi.org/10.1038/s41586-022-04785-z
}
