% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_postnatal_ultra}
\alias{make_postnatal_ultra}
\title{ultrametric making workflow for postnatal part of the tree (takes output of correct_branches as input)
Get the root node (in a rooted tree, root is node Ntip + 1)
use subtract_mutations function to remove prenatal section of tree. Start from the root and remove 79 mutations
perform scaling of post-natal tree to make ultrametric
make age corrected}
\usage{
make_postnatal_ultra(tree_input, mutload_birth, age_at_sampling)
}
\description{
ultrametric making workflow for postnatal part of the tree (takes output of correct_branches as input)
Get the root node (in a rooted tree, root is node Ntip + 1)
use subtract_mutations function to remove prenatal section of tree. Start from the root and remove 79 mutations
perform scaling of post-natal tree to make ultrametric
make age corrected
}
