% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_prenatal_ultra}
\alias{make_prenatal_ultra}
\title{function to create ultrametric prenatal tree
first retrieve prenatal tree
make prenatal tree ultrametric}
\usage{
make_prenatal_ultra(raw_tree, mutations_at_birth)
}
\description{
function to create ultrametric prenatal tree
first retrieve prenatal tree
make prenatal tree ultrametric
}
