% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_prenatal_postnatal}
\alias{merge_prenatal_postnatal}
\title{function to merge the postnatal and prenatal tree
retrieve lengths to be added to postnatal tree
add lengths to postnatal tree
also add this data to the phylo object}
\usage{
merge_prenatal_postnatal(prenatal_tree, postnatal_tree)
}
\description{
function to merge the postnatal and prenatal tree
retrieve lengths to be added to postnatal tree
add lengths to postnatal tree
also add this data to the phylo object
}
