% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perform_age_scaling}
\alias{perform_age_scaling}
\title{from here on Lucca functions again
age scaling of ultrametric trees (ultrametric is initially scaled from 0-1)
perform age scaling and store branch lengths in the data layer again}
\usage{
perform_age_scaling(tree_obj, age_in_years)
}
\description{
from here on Lucca functions again
age scaling of ultrametric trees (ultrametric is initially scaled from 0-1)
perform age scaling and store branch lengths in the data layer again
}
