% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subtract_mutations}
\alias{subtract_mutations}
\title{ultrametric tree making step 1: remove prenatal region of tree
Function to recursively subtract mutations as we traverse the tree from the root >> remove prenatal section of the tree
use find root node as initial input for 'node' and the mut load at birth as initial 'remaining_mutations'
Get the edges that descend from the current node
Get the length of edge
If the edge length is shorter than or equal to the 'remaining mutations', set the whole edge to length 0
If the edge length is longer than the remaining mutations, subtract the 'remaining mutations'
Recursively process the child node (if there are remaining mutations)}
\usage{
subtract_mutations(tree, node, remaining_mutations)
}
\description{
ultrametric tree making step 1: remove prenatal region of tree
Function to recursively subtract mutations as we traverse the tree from the root >> remove prenatal section of the tree
use find root node as initial input for 'node' and the mut load at birth as initial 'remaining_mutations'
Get the edges that descend from the current node
Get the length of edge
If the edge length is shorter than or equal to the 'remaining mutations', set the whole edge to length 0
If the edge length is longer than the remaining mutations, subtract the 'remaining mutations'
Recursively process the child node (if there are remaining mutations)
}
