% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traverse_and_cut}
\alias{traverse_and_cut}
\title{ultrametric tree making step 2: isolate only prenatal region of tree
Function to recursively traverse the tree and cut the max distance traversed to 'remaining mutations' (start at mutation load at birth)
Get the edges that descend from the current node
Get the length of edge
If the edge length is shorter than or equal to the 'remaining mutations', keep whole edge length
update remaining mutations to subtract the edge length of this edge
Recursively process the child node (if there are remaining mutations)
If the edge length is longer than the remaining mutations, set to 'remaining mutations'
set length of descending branches to 0 as these are all cut off
first get the index of all the nodes and tips downstream of current node
get descending nodes
set length of the edges belonging to those nodes to 0}
\usage{
traverse_and_cut(tree, node, remaining_mutations = 79)
}
\description{
ultrametric tree making step 2: isolate only prenatal region of tree
Function to recursively traverse the tree and cut the max distance traversed to 'remaining mutations' (start at mutation load at birth)
Get the edges that descend from the current node
Get the length of edge
If the edge length is shorter than or equal to the 'remaining mutations', keep whole edge length
update remaining mutations to subtract the edge length of this edge
Recursively process the child node (if there are remaining mutations)
If the edge length is longer than the remaining mutations, set to 'remaining mutations'
set length of descending branches to 0 as these are all cut off
first get the index of all the nodes and tips downstream of current node
get descending nodes
set length of the edges belonging to those nodes to 0
}
